support migration
Migration to support both VueJS 2 & version 3.  To handle version 3 properly, had to produce a createInstance function instead of a constructor and introduced the ability to convert the supplied object into a VueJS 3 app/Vue through the method toVue().  Warning, this is only supported on a single object, not on the arrays.